requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
enum-compat==0.0.3
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.13.1
fonttools==4.57.0
fsspec==2024.6.1
grpcio==1.71.0
h11==0.14.0
idna==3.10
Jinja2==3.1.3
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==24.2
pillow==11.0.0
protobuf==6.30.2
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.5.1+cu124
torch-model-archiver==0.12.0
torchaudio==2.5.1+cu124
torchserve==0.12.0
torchvision==0.20.1+cu124
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
Werkzeug==3.1.3
